ir related aggregate types from aggregate to grouping ir module
